Working Win Screen
Nach einer bestimmten Zeit poppet der Win Screen auf, auf dem man das Level neu starten oder die Level Auswahl aufrufen kann
