+++ failedRAM/Assets/Scripte/Anderes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae480100dbaa580439e1065cabe9e294
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
