+++ failedRAM/Assets/Scripte/Anderes/ObjektPulser.cs
@@ -1,0 +1,38 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ObjektPulser : MonoBehaviour
+{
+    [SerializeField] bool _useTestBeat;
+    [SerializeField] float _pulseSize = 1.15f;
+    [SerializeField] float _returnSpeed = 5f;
+    private Vector3 _startSize;
+
+    private void Start()
+    {
+        _startSize = transform.localScale;
+        if (_useTestBeat)
+        {
+            StartCoroutine(TestBeat());
+        }
+    }
+
+    private void Update()
+    {
+        transform.localScale = Vector3.Lerp(transform.localScale, _startSize, Time.deltaTime * _returnSpeed);
+    }
+
+    public void Pulse()
+    {
+        transform.localScale = _startSize * _pulseSize;
+    }
+    IEnumerator TestBeat()
+    {
+        while (true)
+        {
+            yield return new WaitForSeconds(1f);
+            Pulse();
+        }
+    }
+}
