--- failedRAM/Assets/Scripte/HealthBar.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-public class HealthBar : MonoBehaviour
-{
-    [SerializeField]
-    Slider slider;
-
-    public void SetMaxHealth(int health)
-    {
-        slider.maxValue = health;
-        slider.value = health;
-    }
-
-    public void SetHealth(int health) 
-    {
-        slider.value = health;
-    }
-  
-}
