+++ failedRAM/Assets/Scripte/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca297c37d89daa04bafe6cdfe570a04a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
