+++ failedRAM/Assets/Scripte/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cc6b06d199f7bdf42ad5ad3e0401fc78
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
