+++ failedRAM/Assets/Scripte/UI/WinScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 841bffbf922ab7449bd86b45ec5bd7ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
